Bis Gegner.h alles mit Kommentaren beschrieben den Rest mache ich noch, aber net jetzt

+++ Beispielprojekt/ButtonHandler.cpp
@@ -8,14 +8,14 @@ using namespace std;
 vector<vector<Kachel>> Maustaste_Losgelassen(vector<vector<Kachel>> arrayKachel, int x_maus, int y_maus)
 {
 
-	int iterator_ze = 1;
-	int iterator_sp = 1;
+	double iterator_ze = 1;
+	double iterator_sp = 1;
 
 	for (iterator_ze = 1; iterator_ze < arrayKachel.size() + 1; iterator_ze++)
 	{
 		for (iterator_sp = 1; iterator_sp < arrayKachel.at(0).size() + 1; iterator_sp++)
 		{
-			arrayKachel[int(iterator_ze - 1)][int(iterator_sp - 1)].set_change(false);
+			arrayKachel[(iterator_ze - 1)][int(iterator_sp - 1)].set_change(false);
 		}
 	}
 	return arrayKachel;
